Fallback to generic code on JULIA_CPU_TARGET
In FASTX.jl issue #101, it was discovered that pkgimages in 1.9 do not play well
with ScanByte and precompilation.
During precompilation, ScanByte detects the CPU features and produce code making
use of these features. However, pkgimages cannot handle this if JULIA_CPU_TARGET
is set to "generic".

Fix this by, if JULIA_CPU_TARGET is set, fallback to generic code.